New route rest calls first commit

+++ app/src/main/java/hr/fer/trackmyroute/data/model/Route.kt
@@ -9,7 +9,7 @@ data class Route(
     var name: String = "",
     // problem with date format when converting back to json (using string cause of that)
     var date: String = "",
-    var duration: Long = -1,
+    var duration: Double = -1.0,
     var speed: Double = -1.0,
     var distance: Double = -1.0
 ) : Serializable
